workspace.requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiokafka==0.12.0
aiosignal==1.3.1
alembic==1.14.0
alibi-detect==0.12.0
annotated-types==0.7.0
ansible==11.0.0
ansible-compat==24.10.0
ansible-core==2.18.0
ansible-creator==24.11.0
ansible-lint==24.10.0
anyio==4.6.2.post1
apache-airflow-client==2.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-property==0.2.2
async-timeout==5.0.1
attrs==23.2.0
black==24.10.0
blinker==1.9.0
blis==0.7.11
boto3==1.35.57
botocore==1.35.57
bracex==2.5.post1
Brotli==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.8
databricks-sdk==0.36.0
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
durationpy==0.9
fastapi==0.115.5
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gevent==24.11.1
geventhttpclient==2.0.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
googleapis-common-protos==1.66.0
graphene==3.4.1
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.67.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.7
langchain-core==0.3.17
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langcodes==3.4.1
langsmith==0.1.142
language_data==1.2.0
lazy_loader==0.4
llvmlite==0.42.0
Mako==1.3.6
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
minio==7.2.11
mlflow==2.17.2
mlflow-skinny==2.17.2
mlserver==1.3.5
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai==1.54.4
opencv-python==4.10.0.84
opentelemetry-api==1.28.2
opentelemetry-distro==0.49b2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
preshed==3.0.9
prometheus_client==0.21.0
propcache==0.2.0
protobuf==5.28.3
py-grpc-prometheus==0.8.0
py4j==0.10.9.7
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.10.0
PyNaCl==1.5.0
pyparsing==3.2.0
pyre-extensions==0.0.31
pyspark==3.5.3
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-keycloak==4.7.2
python-multipart==0.0.17
python-rapidjson==1.20
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resolvelib==1.0.1
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
SQLAlchemy==2.0.36
sqlparse==0.5.1
srsly==2.4.8
starlette==0.41.2
starlette-exporter==0.17.1
subprocess-tee==0.4.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
torchx==0.7.0
tqdm==4.67.0
transformers==4.46.2
triton==3.1.0
tritonclient==2.51.0
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.32.0
uvloop==0.21.0
wasabi==1.1.3
wcmatch==10.0
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.16.0
yamllint==1.35.1
yarl==1.17.1
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.1
